+++ src/main/java/janggi/service/JanggiService.java
@@ -3,10 +3,10 @@ package janggi.service;
 import janggi.GameState;
 import janggi.board.Board;
 import janggi.board.BoardGenerator;
-import janggi.dao.GameDao;
-import janggi.dao.PieceDao;
 import janggi.dao.entity.GameEntity;
 import janggi.dao.entity.PieceEntity;
+import janggi.dao.game.GameDao;
+import janggi.dao.piece.PieceDao;
 import janggi.piece.Piece;
 import janggi.piece.Team;
 import janggi.position.Position;
